requirements.txt
pandas==2.2.2
numpy==1.26.4

scikit-learn==1.5.1
joblib==1.4.2

torch>=2.2.0
transformers>=4.41.0
tokenizers>=0.15.0

selenium==4.22.0
beautifulsoup4==4.12.3

python-dotenv==1.0.1

streamlit==1.37.0
plotly==5.22.0

tqdm==4.66.4